Fix integer overflow causing seekg errors in vtkNek5000Reader

+++ IO/Parallel/vtkNek5000Reader.h
@@ -182,7 +182,7 @@ protected:
   int myNumBlockReads;
   int* myBlockIDs;
   int* proc_numBlocks;
-  int* myBlockPositions;
+  long* myBlockPositions;
   int NumberOfTimeSteps;
   double TimeValue;
   int TimeStepRange[2];
